Use a tooltip for the Promo-blocking upsell
The previous implementation was pretty confusing, with it being
possible to have multiple stops on the slider selected.

With this commit, the upsell is instead a button that is not one of
the stops on the slider, and triggers a tooltip explaining that it
is only available to Premium subscribers.

It's still somewhat confusing, in that you can tab first to the
button, and then to the slider thumb that you control with the
arrow keys, even though they both appear to be part of the same
slider, but it's a lot better than it was.